refactor: folder structure and studio import alias
