--- src/app/pages/Home/components/index.ts
@@ -1,3 +1,0 @@
-export { BotDetail } from './BotDetail';
-export { ContentTab } from './ContentTab';
-export { SidebarContent } from './SidebarContent';
