+++ src/app/pages/Home/containers/Appbar/components/ContentTab/styles.ts
