+++ src/app/pages/Home/containers/Appbar/components/index.ts
@@ -1,0 +1,1 @@
+export * from './ContentTab';
