+++ src/app/pages/Home/containers/HomeContent/components/BotDetail/BotDetail.tsx
