+++ src/app/pages/Home/containers/HomeContent/components/BotDetail/helpers.ts
