+++ src/app/pages/Home/containers/HomeContent/components/BotDetail/index.ts
