+++ src/app/pages/Home/containers/HomeContent/components/BotDetail/styles.ts
