+++ src/app/pages/Home/containers/HomeContent/components/index.ts
@@ -1,0 +1,1 @@
+export * from './BotDetail';
