+++ src/app/pages/Home/containers/Sidebar/components/SidebarContent/SidebarContent.tsx
