+++ src/app/pages/Home/containers/Sidebar/components/SidebarContent/helpers.tsx
