+++ src/app/pages/Home/containers/Sidebar/components/SidebarContent/index.ts
