+++ src/app/pages/Home/containers/Sidebar/components/SidebarContent/styles.ts
