+++ src/app/pages/Home/containers/Sidebar/components/index.ts
@@ -1,0 +1,1 @@
+export * from './SidebarContent';
