--- src/app/pages/Studio/components/ContextMenu/index.ts
@@ -1,1 +1,0 @@
-export { ContextMenu } from './ContextMenu';
