--- src/app/pages/Studio/components/index.ts
@@ -1,6 +1,0 @@
-export { Flows } from './Flows';
-export { EndNode } from './EndNode';
-export { StartNode } from './StartNode';
-export { ContextMenu } from './ContextMenu';
-export { StandardNode } from './StandardNode';
-export { ExplorerHandle } from './ExplorerHandle';
