+++ src/app/pages/Studio/containers/Canvas/Canvas.tsx
@@ -5,12 +5,12 @@ import ReactFlow, {
 } from 'reactflow';
 
 import { colors } from '@/styles';
+import { CanvasProvider } from '@studio/contexts';
 
 import { nodeTypes } from './constants';
 import { usePrepareHook } from './helpers';
 
-import { ContextMenu } from '../../components';
-import { CanvasProvider } from '../../contexts';
+import { ContextMenu } from './components';
 
 export const Canvas = () => {
   const { edges, nodes, onConnect, onEdgesChange, onNodesChange, onShowMenu } =
