+++ src/app/pages/Studio/containers/Canvas/components/ContextMenu/index.ts
@@ -1,0 +1,2 @@
+export * from './constants';
+export * from './ContextMenu';
