+++ src/app/pages/Studio/containers/Canvas/components/EndNode/EndNode.tsx
