+++ src/app/pages/Studio/containers/Canvas/components/EndNode/index.ts
