+++ src/app/pages/Studio/containers/Canvas/components/EndNode/styles.ts
