+++ src/app/pages/Studio/containers/Canvas/components/Handle/index.ts
