+++ src/app/pages/Studio/containers/Canvas/components/Handle/styles.css
