+++ src/app/pages/Studio/containers/Canvas/components/StandardNode/StandardNode.tsx
