+++ src/app/pages/Studio/containers/Canvas/components/StandardNode/styles.ts
