+++ src/app/pages/Studio/containers/Canvas/components/StartNode/StartNode.tsx
