+++ src/app/pages/Studio/containers/Canvas/components/StartNode/index.ts
