+++ src/app/pages/Studio/containers/Canvas/components/StartNode/styles.ts
