+++ src/app/pages/Studio/containers/Canvas/constants.ts
@@ -1,6 +1,6 @@
 import { NodeTypeEnum } from '@/enums';
 
-import { StartNode, StandardNode, EndNode } from '../../components';
+import { EndNode, StartNode, StandardNode } from './components';
 
 export const nodeTypes = {
   [NodeTypeEnum.END]: EndNode,
