+++ src/app/pages/Studio/containers/Explorer/components/ExplorerHandle/styles.ts
