+++ src/app/pages/Studio/containers/Explorer/helpers.tsx
@@ -3,6 +3,11 @@ import { useTranslation } from 'react-i18next';
 import { SyntheticEvent, useMemo, useRef } from 'react';
 
 import {
+  ExplorerType,
+  MAX_EXPLORER_WIDTH,
+  MIN_EXPLORER_WIDTH,
+} from '@studio/constants';
+import {
   setWidth,
   setDragging,
   setExplorer,
@@ -13,12 +18,6 @@ import {
 import { useAppDispatch, useAppSelector, useGetMyBots } from '@/hooks';
 
 import { messages } from './messages';
-
-import {
-  ExplorerType,
-  MAX_EXPLORER_WIDTH,
-  MIN_EXPLORER_WIDTH,
-} from '../../constants';
 
 export const usePrepareHook = () => {
   const { id } = useParams();
