+++ tsconfig.json
@@ -16,7 +16,8 @@
     "noEmit": true,
     "jsx": "react-jsx",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@studio/*": ["./src/app/pages/Studio/*"]
     }
   },
   "include": ["src"]
